# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Frames API
  version: 0.1.0
servers:
  - url: http://localhost:3001
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /admin/users/{id}:
    get:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Administrator user detail }
...
  /auth/register:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, displayName, email, password]
              properties:
                username: { type: string }
                displayName: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "201":
          description: Created user and tokens
  /auth/login:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Authenticated user and tokens
...
  /users/{username}:
    get:
      parameters:
        - in: path
          name: username
          required: true
          schema: { type: string }
      responses:
        "200": { description: User profile }
  /users/{id}/friend-request:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "201": { description: Friend request }
...
  /posts/{id}:
    get:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Post detail }
    delete:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted post }
  /posts/{id}/reactions:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "201": { description: Reaction }
  /posts/{id}/comments:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "201": { description: Comment }
...
  /archive/daily/{date}:
    get:
      parameters:
        - in: path
          name: date
          required: true
          schema: { type: string, format: date }
      responses:
        "200": { description: Daily frame }
  /memories/monthly/{year}/{month}:
    get:
      parameters:
        - in: path
          name: year
          required: true
          schema: { type: integer }
        - in: path
          name: month
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Monthly collage }
  /memories/yearly/{year}:
    get:
      parameters:
        - in: path
          name: year
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Yearbook }
...